Changed configuration file references for Exchange API config.
Starting in version 2.22.0 of the Exchange API the container is now
responsible for the creation of /etc/horizon/exchange/config.json. The
Dockerfile uses /etc/horizon/exchange/exchange-api.tmpl to build the
final configuration file for the application. This change is not
backwards compatible with older versions of the exchange.

Minimum Exchange version bumped from version 2.20.0 to 2.22.0.